+++ test/cli/job.js
@@ -33,7 +33,7 @@ buster.testCase('cli - job', {
     job.create(this.mockArgsCb)('somejob', 'config.xml');
   },
   'read - should log job status with correct color and health reports': function () {
-    this.mockConsole.expects('log').once().withExactArgs('%s | %s', 'somejob', 'OK'.blue);
+    this.mockConsole.expects('log').once().withExactArgs('%s | %s', 'somejob', 'ok'.blue);
     this.mockConsole.expects('log').once().withExactArgs(' - %s', 'somereport1');
     this.mockConsole.expects('log').once().withExactArgs(' - %s', 'somereport2');
     this.mockProcess.expects('exit').once().withExactArgs(0);
@@ -53,7 +53,7 @@ buster.testCase('cli - job', {
     job.read(this.mockArgsCb)('somejob');
   },
   'read - should log status as-is in grey color when color is status value': function () {
-    this.mockConsole.expects('log').once().withExactArgs('%s | %s', 'somejob', 'NOTBUILT'.grey);
+    this.mockConsole.expects('log').once().withExactArgs('%s | %s', 'somejob', 'notbuilt'.grey);
     this.mockConsole.expects('log').once().withExactArgs(' - %s', 'somereport1');
     this.mockConsole.expects('log').once().withExactArgs(' - %s', 'somereport2');
     this.mockProcess.expects('exit').once().withExactArgs(0);