+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 0.3.0-pre
 * Reorganise code structure
 * Move buildlight support to http://github.com/cliffano/nestor-buildlight
+* Move ninjablocks support to http://github.com/cliffano/nestor-ninjablocks
 
 ### 0.2.4
 * Add create-view, update-view, and fetch-view-config commands