+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-### 0.3.0-pre
+### 0.3.1-pre
+*
+
+### 0.3.0
 * Modify Jenkins module to pass Jenkins REST API as-is, update method names (NOTE: backward incompatible with <= 0.2.x)
 * Reorganise code structure
 * Move buildlight support to http://github.com/cliffano/nestor-buildlight